refactor: migration to bevy_fast_tilemap

+++ src/utils.rs
@@ -1,3 +1,3 @@
-pub fn get_continual_coord(n: i32, max: u32) -> u32 {
-    (n).rem_euclid(max as i32) as u32
+pub fn get_continual_coord(n: i64, max: u32) -> u32 {
+    (n).rem_euclid(max as i64) as u32
 }
